refactor(Site): :recycle: Re-organize site components
